精简抽数计算 Go 实现
将仓库翻页结束控制交给 Pipeline max_hit，移除抽数计算的自定义识别。

固定抽数公式常量，减少 Pipeline 参数解析，并合并 pullcount 阶段状态文件。
