+++ agent/go-service/pullcount/register.go
@@ -5,5 +5,4 @@ import maa "github.com/MaaXYZ/maa-framework-go/v4"
 // Register registers pull count calculator custom components.
 func Register() {
 	maa.AgentServerRegisterCustomAction("PullCountCalculatorAction", &Action{})
-	maa.AgentServerRegisterCustomRecognition("PullCountCalculatorRecognition", &Recognition{})
 }
