+++ tools/schema/custom.recognition.schema.json
@@ -17,7 +17,6 @@
                 "MapTrackerAssertLocation",
                 "MapTrackerBigMapInfer",
                 "MapTrackerInfer",
-                "PullCountCalculatorRecognition",
                 "PuzzleRecognition",
                 "SellProductNormalizedItemMatch",
                 "VisitFriendsMenuScanDetailAssistRecognition",
